refactor: merge modo-i18n into modo core behind 'i18n' feature flag
Move i18n source files into modo/src/i18n/, t\! macro into modo-macros.
Remove modo-i18n, modo-i18n-macros, modo-templates, and
modo-templates-macros from workspace and delete their directories.
Workspace reduced from 17 to 12 crates.
