+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["modo", "modo-macros", "modo-db", "modo-db-macros", "modo-email", "modo-i18n", "modo-i18n-macros", "modo-jobs", "modo-jobs-macros", "modo-session", "modo-auth", "modo-tenant", "modo-upload", "modo-upload-macros", "modo-templates", "modo-templates-macros", "examples/*"]
+members = ["modo", "modo-macros", "modo-db", "modo-db-macros", "modo-email", "modo-jobs", "modo-jobs-macros", "modo-session", "modo-auth", "modo-tenant", "modo-upload", "modo-upload-macros", "examples/*"]
 
 [workspace.package]
 license = "Apache-2.0"
