--- modo-i18n-macros/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "modo-i18n-macros"
-version = "0.1.0"
-edition = "2024"
-license.workspace = true
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = { version = "2", features = ["full", "extra-traits"] }
-quote = "1"
-proc-macro2 = "1"
