--- modo-i18n/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "modo-i18n"
-version = "0.1.0"
-edition = "2024"
-license.workspace = true
-
-[dependencies]
-modo = { path = "../modo" }
-modo-i18n-macros = { path = "../modo-i18n-macros" }
-
-axum = "0.8"
-axum-extra = { version = "0.10", features = ["cookie"] }
-http = "1"
-tower = { version = "0.5", features = ["util"] }
-serde = { version = "1", features = ["derive"] }
-serde_yaml_ng = "0.10"
-futures-util = "0.3"
-tracing = "0.1"
-
-modo-templates = { path = "../modo-templates", optional = true }
-minijinja = { version = "2", optional = true }
-
-[features]
-default = []
-templates = ["dep:modo-templates", "dep:minijinja"]
-
-[dev-dependencies]
-minijinja = "2"
-tokio = { version = "1", features = ["full"] }
-tower = { version = "0.5", features = ["util"] }
-http = "1"
