+++ modo-macros/src/lib.rs
@@ -7,6 +7,7 @@ mod middleware;
 mod module;
 mod sanitize;
 mod validate;
+mod t_macro;
 mod view;
 
 /// Attribute macro for declaring HTTP handlers with auto-registration.
@@ -53,6 +54,14 @@ pub fn derive_sanitize(input: TokenStream) -> TokenStream {
 #[proc_macro_derive(Validate, attributes(validate))]
 pub fn derive_validate(input: TokenStream) -> TokenStream {
     validate::expand(input.into())
+        .unwrap_or_else(|e| e.to_compile_error())
+        .into()
+}
+
+/// Translate a key with optional named variables.
+#[proc_macro]
+pub fn t(input: TokenStream) -> TokenStream {
+    t_macro::expand(input.into())
         .unwrap_or_else(|e| e.to_compile_error())
         .into()
 }
