+++ modo-macros/src/t_macro.rs
@@ -1,4 +1,3 @@
-use proc_macro::TokenStream;
 use quote::quote;
 use syn::parse::{Parse, ParseStream};
 use syn::{Expr, Ident, LitStr, Token};
@@ -35,21 +34,7 @@ impl Parse for TInput {
     }
 }
 
-/// Translate a key with optional named variables.
-///
-/// Usage:
-/// - `t!(i18n, "key")`
-/// - `t!(i18n, "key", name = expr)`
-/// - `t!(i18n, "key", count = expr)` — triggers plural
-#[proc_macro]
-pub fn t(input: TokenStream) -> TokenStream {
-    match t_impl(input.into()) {
-        Ok(tokens) => tokens.into(),
-        Err(err) => err.to_compile_error().into(),
-    }
-}
-
-fn t_impl(input: proc_macro2::TokenStream) -> syn::Result<proc_macro2::TokenStream> {
+pub fn expand(input: proc_macro2::TokenStream) -> syn::Result<proc_macro2::TokenStream> {
     let input = syn::parse2::<TInput>(input)?;
     let i18n = &input.i18n_expr;
     let key = &input.key;
