--- modo-templates/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "modo-templates"
-version = "0.1.0"
-edition = "2024"
-license.workspace = true
-
-[dependencies]
-modo-templates-macros = { path = "../modo-templates-macros" }
-
-minijinja = { version = "2", features = ["loader"] }
-serde = { version = "1", features = ["derive"] }
-serde_yaml_ng = "0.10"
-axum = "0.8"
-http = "1"
-tower = { version = "0.5", features = ["util"] }
-futures-util = "0.3"
-tracing = "0.1"
-
-[dev-dependencies]
-tokio = { version = "1", features = ["full"] }
-tower = { version = "0.5", features = ["util"] }
-http = "1"
-serde = { version = "1", features = ["derive"] }
