+++ modo/Cargo.toml
@@ -62,6 +62,7 @@ mime_guess = { version = "2", optional = true }
 default = []
 templates = ["dep:minijinja", "dep:futures-util", "dep:http"]
 csrf = ["dep:rand", "dep:hmac", "dep:sha2", "dep:subtle", "dep:form_urlencoded", "dep:http"]
+i18n = ["dep:futures-util", "dep:http"]
 static-fs = ["tower-http/fs"]
 static-embed = ["dep:rust-embed", "dep:mime_guess", "modo-macros/static-embed"]
