+++ modo/src/i18n/entry.rs
